bidsdir:
Traverse main directory and harvest paths and file names
